refactor(mesh): improved type safety
Renamed `render` to `draw` for consistency across mesh classes. Replaced raw pointer casts with safer `reinterpret_cast` and `static_cast`. Updated loop variable types to `size_t` for better compatibility and added `const` qualifiers to improve immutability where applicable.

+++ src/renderer/mesh/texture.cpp
@@ -44,7 +44,7 @@ unsigned int texture::loadCubemap(std::vector<std::string>& faces) {
     glBindTexture(GL_TEXTURE_CUBE_MAP, textureID);
 
     int width, height, depth;
-    for (unsigned int i = 0; i < faces.size(); i++) {
+    for (size_t i = 0; i < faces.size(); i++) {
         unsigned char* data = stbi_load(faces[i].c_str(), &width, &height, &depth, 0);
 
         if (!data) {
